+++ RELEASES.md
@@ -2,11 +2,11 @@
 
 ## [v1.12.3] (pending)
 
-- Added a health check that alerts if a primary network validator has no nodes connected to it. Runs a configurable time after bootstrapping or 20 minutes by default.
+- Extended the network health check by also alerting if a primary network validator has no nodes connected to it. Runs a configurable time after startup or 10 minutes by default.
 
 ### Configs
--  How long after bootstrapping the aforementioned health check runs can be configued via:
-`--no-ingress-validator-connection-timeout`
+-  How long after startup the aforementioned health check runs can be configured via:
+`--network-no-ingress-connections-grace-period`
 
 
 ## [v1.12.2](https://github.com/ava-labs/avalanchego/releases/tag/v1.12.2)